requirements.txt
streamlit==1.26.0
streamlit-option-menu==0.5.4
requests==2.31.0
plotly==5.16.1
Flask==2.3.6
psycopg2-binary==2.9.9
web3==6.10.0
easyocr==1.6.2
opencv-python-headless==4.8.1.78
numpy==1.26.2
qrcode==7.4.2
PyJWT==2.8.2
